pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "audhd-agents"
version = "0.1.0"
description = "Multi-agent orchestration system designed for AuDHD cognition."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "google-genai>=1.67.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["adapters*", "runtime*", "scripts*", "graphs*", "capabilities*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
